+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### Unreleased
 
+- Added change password functionality #5
 - Added logout route #4
 - Added login and register routes #2, #3
 - Added basic npm files, added express as a setup, added hello world functionality #1